+++ src/evaluate.h
@@ -24,8 +24,53 @@
 #include <string>
 
 #include "types.h"
+#include "material.h"
+#include "pawns.h"
 
 class Position;
+
+// Struct EvalInfo contains various information computed and collected
+  // by the evaluation functions.
+  struct EvalInfo {
+
+    // attackedBy[color][piece type] is a bitboard representing all squares
+    // attacked by a given color and piece type (can be also ALL_PIECES).
+    Bitboard attackedBy[COLOR_NB][PIECE_TYPE_NB];
+
+    // attackedBy2[color] are the squares attacked by 2 pieces of a given color,
+    // possibly via x-ray or by one pawn and one piece. Diagonal x-ray through
+    // pawn or squares attacked by 2 pawns are not explicitly added.
+    Bitboard attackedBy2[COLOR_NB];
+
+    // kingRing[color] is the zone around the king which is considered
+    // by the king safety evaluation. This consists of the squares directly
+    // adjacent to the king, and the three (or two, for a king on an edge file)
+    // squares two ranks in front of the king. For instance, if black's king
+    // is on g8, kingRing[BLACK] is a bitboard containing the squares f8, h8,
+    // f7, g7, h7, f6, g6 and h6.
+    Bitboard kingRing[COLOR_NB];
+
+    // kingAttackersCount[color] is the number of pieces of the given color
+    // which attack a square in the kingRing of the enemy king.
+    int kingAttackersCount[COLOR_NB];
+
+    // kingAttackersWeight[color] is the sum of the "weights" of the pieces of the
+    // given color which attack a square in the kingRing of the enemy king. The
+    // weights of the individual piece types are given by the elements in the
+    // KingAttackWeights array.
+    int kingAttackersWeight[COLOR_NB];
+
+    // kingAdjacentZoneAttacksCount[color] is the number of attacks by the given
+    // color to squares directly adjacent to the enemy king. Pieces which attack
+    // more than one square are counted multiple times. For instance, if there is
+    // a white knight on g5 and black's king is on g8, this white knight adds 2
+    // to kingAdjacentZoneAttacksCount[WHITE].
+    int kingAdjacentZoneAttacksCount[COLOR_NB];
+
+    Bitboard pinnedPieces[COLOR_NB];
+    Material::Entry* me;
+    Pawns::Entry* pi;
+  };
 
 namespace Eval {
 
@@ -35,7 +80,9 @@ void init();
 std::string trace(const Position& pos);
 
 template<bool DoTrace = false>
-Value evaluate(const Position& pos);
+Value evaluate(const Position& pos, EvalInfo& ei);
 }
+
+
 
 #endif // #ifndef EVALUATE_H_INCLUDED
